+++ Include/Acidanthera/Library/OcConfigurationLib.h
@@ -459,7 +459,7 @@ typedef enum {
   _(UINT16                      , ProcessorType      ,     , 0                                            , () )                   \
   _(UINT8                       , Rom                , [6] , {0}                                          , () )                   \
   _(BOOLEAN                     , SpoofVendor        ,     , FALSE                                        , () )                   \
-  _(BOOLEAN                     , AdviseWindows      ,     , FALSE                                        , () )                   \
+  _(BOOLEAN                     , AdviseFeatures     ,     , FALSE                                        , () )                   \
   _(BOOLEAN                     , MaxBIOSVersion     ,     , FALSE                                        , () )
   OC_DECLARE (OC_PLATFORM_GENERIC_CONFIG)