+++ Library/OcConfigurationLib/OcConfigurationLib.c
@@ -556,7 +556,7 @@ mPlatformConfigurationDataHubSchema[] = {
 STATIC
 OC_SCHEMA
 mPlatformConfigurationGenericSchema[] = {
-  OC_SCHEMA_BOOLEAN_IN ("AdviseWindows",      OC_GLOBAL_CONFIG, PlatformInfo.Generic.AdviseWindows),
+  OC_SCHEMA_BOOLEAN_IN ("AdviseFeatures",     OC_GLOBAL_CONFIG, PlatformInfo.Generic.AdviseFeatures),
   OC_SCHEMA_STRING_IN  ("MLB",                OC_GLOBAL_CONFIG, PlatformInfo.Generic.Mlb),
   OC_SCHEMA_BOOLEAN_IN ("MaxBIOSVersion",     OC_GLOBAL_CONFIG, PlatformInfo.Generic.MaxBIOSVersion),
   OC_SCHEMA_INTEGER_IN ("ProcessorType",      OC_GLOBAL_CONFIG, PlatformInfo.Generic.ProcessorType),